add default deposit app validation to node and client
