+++ modules/node/src/cfCore/cfCore.provider.ts
@@ -1,5 +1,6 @@
+import { generateValidationMiddleware } from "@connext/apps";
 import { MessagingService } from "@connext/messaging";
-import { CF_PATH, ConnextNodeStorePrefix, IMessagingService } from "@connext/types";
+import { CF_PATH, ConnextNodeStorePrefix, IMessagingService, Opcode, ContractAddresses } from "@connext/types";
 import { Provider } from "@nestjs/common";
 import { fromMnemonic } from "ethers/utils/hdnode";
 
@@ -29,10 +30,11 @@ export const cfCoreProviderFactory: Provider = {
 
     // test that provider works
     const { chainId, name: networkName } = await config.getEthNetwork();
+    const contractAddresses = await config.getContractAddresses();
     const cfCore = await CFCore.create(
       messaging as IMessagingService, // TODO: FIX
       store,
-      await config.getContractAddresses(),
+      contractAddresses,
       { STORE_KEY_PREFIX: ConnextNodeStorePrefix },
       provider,
       { acquireLock: lockService.lockedOperation.bind(lockService) },
@@ -43,6 +45,14 @@ export const cfCoreProviderFactory: Provider = {
       },
       undefined,
       log.newContext("CFCore"),
+    );
+    // inject any default middlewares
+    cfCore.injectMiddleware(
+      Opcode.OP_VALIDATE,
+      await generateValidationMiddleware({
+        ...contractAddresses,
+        provider,
+      } as ContractAddresses),
     );
     const signerAddr = await cfCore.signerAddress();
     const balance = (await provider.getBalance(signerAddr)).toString();
