+++ modules/types/src/protocol.ts
@@ -75,6 +75,24 @@ type UpdateProtocolParams = {
 ////////////////////////////////////////
 // exports
 
+export enum Opcode {
+  // Middleware hook to send a ProtocolMessage to a peer.
+  IO_SEND,
+  // Middleware hook to both send and wait for a response from a ProtocolMessage
+  IO_SEND_AND_WAIT,
+  // Requests a signature on the hash of previously generated EthereumCommitments.
+  OP_SIGN,
+  // Middleware hook to write the app instances to store.
+  PERSIST_APP_INSTANCE,
+  // Called at the end of execution before the return value to store a commitment
+  PERSIST_COMMITMENT,
+  // Middleware hook to write the state channel to store. Used to lock channel between protocols.
+  PERSIST_STATE_CHANNEL,
+  // Middleware hook to validate state transitions in protocol. Called before
+  // `computeStateTransition` and registered using `injectMiddleware`
+  OP_VALIDATE,
+}
+
 export const ProtocolNames = enumify({
   install: "install",
   propose: "propose",
